requirements.txt
 
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.15
pymysql==1.1.0
pydantic==1.10.7
alembic==1.11.1
logging==0.4.9.6
pydantic[email]
PyJWT
bcrypt